// src/features/editor/types/task.ts
export enum TaskType {
  FORM_INPUTS = "FORM_INPUTS",
  ADDITION_OPREATION = "ADDITION_OPREATION",
  SUBTRACTION_OPREATION = "SUBTRACTION_OPREATION",
  MULTIPLICATION_OPREATION = "MULTIPLICATION_OPREATION",
  DIVISION_OPREATION = "DIVISION_OPREATION",
  PERCENTAGE_OPREATION = "PERCENTAGE_OPREATION",
  NUMBER_TO_STRING_OPREATION = "NUMBER_TO_STRING_OPREATION",
  LETTERCASE_OPREATION = "LETTERCASE_OPREATION",
  NUMBER_FORMAT_OPREATION = "NUMBER_FORMAT_OPREATION",
  PREFIX_OPREATION = "PREFIX_OPREATION",
  SUFFIX_OPREATION = "SUFFIX_OPREATION",
  DOCUMENT_NODE = "DOCUMENT_NODE",
}

export enum ParentTaskType {
  FORM_NODE = "FORM_NODE",
  OPREATION_NODE = "OPREATION_NODE",
  DOCUMENT_NODE = "DOCUMENT_NODE",
}